requirements.txt
pyserial 
pandas 
matplotlib 
numpy